Database model and security strategy
The database used is relational database as we want the data to be structured and will be further used for insights and analysis. The database should be TDE encrypted and the fields in the table having personal identifying information must be encrypted. The data in transmission should be shared through secured channel with no PII leakages. 
Refer the database folder and security_strategy.db files